Hex detection false positive
Hi, if you have something like this: 
```
.
.
^FO605,995^A0N,40,40^FDA1_3^FS
.
.
```

The analyzer will detect "_3" as an hex data, but is not, and when the function ReplaceSpecialChars is called an Exception will be thrown with a message like "The index doesnt exists" and is because the function uses
```
 text.Substring(i + 1, 2)
```
And in this case, the length "2" doesnt exists, because the real length is only 1.

Another case is this "_C3_A9"  where a multple Hex data was provided. The replace was taked separately and the real character was not replaced (in this case was é).

So first I created a function to verify if the substring exists, in this function if a Exception is thrown, will return false, and the ReplaceSpecialChars would append the text without trying to replace Hex.

Also for the second case, I verified if substring exists and if third character is "_".